xlnx: rename assists/xlnx_versal_power.py to assists/xlnx/power.py
xlnx_versal_power.py belongs with subsystem and protection modules under
assists/xlnx/* since it is a helper file created for the same purpose.

Move it accordingly and remove references to it if not used in a plugin.

cdo_topology.py does not make use of it, so do not import it there.
