+++ assists/xlnx/power.py
